YaFSDP support inference mode
This PR adds basic support for inference mode to YaFSDP.  In th inference case, there is no need to create buffers for grads and create a high-precision copy of model weights.

This can be useful when the model uses the ZeRO-3 mode with additional parallelism strategies (TP/SP/PP) without training.
60a47ae2ed2cf8403811727245fb66f0c426bde9