Fix excessive memory consumption for YCCL AllGather inputs in YaFSDP
Currently, the output buffers for all-gather are over-allocated so that they are valid for all-gathers using the whole input buffer (i.e., the output size is the *total sum of all parameter shards* multiplied by the number of ranks). For ZeRO-2-like setups, where we might allocate tens or thousands of all-gather output buffers, this is prohibitively large and results in OOMs. Fix this by making the YCCL output buffers the same size as the NCCL ones (i.e., the *maximum parameter shard size* multiplied by the number of ranks).
commit_hash:4117d70970081e4dfa69293ff3cf5266764ab82e